# remove per-feature logging to improve performance issue#685