--- a/fiona/compat.py
+++ b/fiona/compat.py
@@ -1,5 +1,5 @@
-import sys
 import collections
+import sys
 
 try:
     from collections import OrderedDict
@@ -24,3 +24,11 @@ else:
 # More specifically, rasterio has a CRS() class that subclasses UserDict()
 # In Python 2 UserDict() is in its own module and does not subclass Mapping()
 DICT_TYPES = (dict, Mapping, UserDict)
+
+
+def strencode(instr, encoding="utf-8"):
+    try:
+        instr = instr.encode(encoding)
+    except (UnicodeDecodeError, AttributeError):
+        pass
+    return instr