@snowman2 2019-02-03
Looks like the GDAL trunk is using PROJ.6 now: `configure: error: PROJ 6 symbols not found`