@snowman2 2019-02-08
Another idea I had recently was change:
https://github.com/Toblerity/Fiona/blob/013dec95f3a468459a76d57b60f676b415bda738/fiona/ogrext.pyx#L320

To instead iterate over:
```python
collection.schema['properties']
```
And set the property to use:
```python
set_field_null(cogr_feature, i)
```
If the property does not exist in `feature['properties']`.

This would eliminate the need for this check:

https://github.com/Toblerity/Fiona/blob/013dec95f3a468459a76d57b60f676b415bda738/fiona/ogrext.pyx#L1167

But, I wanted to run the idea by you first.