@snowman2 2019-02-09
> I'm inclined to remove the debug statements that you've commented.

I removed those in a separate commit.

> what would you think about holding off on the schema/property checks for now?

I checked the difference between checking the properties and the alternative method and I was surprised that it didn't make a significant difference. So, I agree that it is an unnecessary change. It makes me wonder if it would be a good idea to do a check and remove all unnecessary log.debug/log.info statements in rasterio & fiona for performance reasons?