@sgillies 2019-02-15
I edited the PR title to reflect our decision to take a one thing at a time approach to optimization.